+++ www/app/api/chat/thought/route.ts
@@ -3,18 +3,19 @@ import {
   getUserData,
   HistoryWithoutResponse,
   Message,
+  user,
+  assistant,
   // parsePrompt,
 } from '@/utils/ai';
 import { honcho } from '@/utils/honcho';
-import { render } from '@/utils/prompts/lib';
 import { thoughtPrompt } from '@/utils/prompts/thought';
-// import { readFileSync } from 'fs';
 import { NextRequest, NextResponse } from 'next/server';
-// import path from 'path';
 
-// const thoughtPrompt = readFileSync(
-//   path.join(process.cwd(), 'utils/prompts/thought.md')
-// );
+const CONTEXT_LIMIT = 10;
+
+interface Metadata {
+  summary?: string;
+}
 
 export async function POST(req: NextRequest) {
   const { message, conversationId } = await req.json();
@@ -25,7 +26,6 @@ export async function POST(req: NextRequest) {
   }
   const { appId, userId } = userData;
 
-  const thoughtHistory: Message[] = [];
   const thoughtIter = await honcho.apps.users.sessions.metamessages.list(
     appId,
     userId,
@@ -36,15 +36,9 @@ export async function POST(req: NextRequest) {
     }
   );
 
-  for await (const metamessage of thoughtIter) {
-    if (metamessage.metadata?.type === 'user') {
-      thoughtHistory.push({ role: 'user', content: metamessage.content });
-    } else {
-      thoughtHistory.push({ role: 'assistant', content: metamessage.content });
-    }
-  }
-
-  const promptMessages = render(thoughtPrompt, thoughtHistory);
+  const thoughtHistory = thoughtIter.items.map(
+    (metamessage) => user`${metamessage.content}`
+  );
 
   const recentResponseMeta = await honcho.apps.users.sessions.metamessages.list(
     appId,
@@ -58,25 +52,22 @@ export async function POST(req: NextRequest) {
     }
   );
 
-  let honchoResponse = 'None';
-  let bloomResponse = 'None';
+  const [recentResponse] = recentResponseMeta.items;
+  const content = recentResponse?.content ?? '';
 
-  const responseMetaList = recentResponseMeta.items;
-  if (responseMetaList.length > 0) {
-    const content = responseMetaList[0].content;
-    honchoResponse =
-      content.match(/<honcho>([^]*?)<\/honcho>/)?.[1]?.trim() ?? 'None';
-    bloomResponse =
-      content.match(/<bloom>([^]*?)<\/bloom>/)?.[1]?.trim() ?? 'None';
-  }
+  const honchoResponse =
+    content.match(/<honcho>([^]*?)<\/honcho>/)?.[1]?.trim() ?? 'None';
+  const bloomResponse =
+    content.match(/<bloom>([^]*?)<\/bloom>/)?.[1]?.trim() ?? 'None';
 
-  const messages = [
-    ...promptMessages,
+  const prompt: Message[] = [
+    ...thoughtPrompt,
+    ...thoughtHistory,
     {
       role: 'user',
       content: `<honcho-response>${honchoResponse}</honcho-response>\n<bloom>${bloomResponse}</bloom>\n${message}`,
     },
-  ] as Message[];
+  ];
 
   const honchoPayload: HistoryWithoutResponse = {
     appId,
@@ -85,7 +76,11 @@ export async function POST(req: NextRequest) {
     userInput: message,
   };
 
-  const stream = await createStream('thought', messages, honchoPayload);
+  console.log('Messages:\n');
+  console.log(prompt);
+  console.log('\n\n\n');
+
+  const stream = await createStream('thought', prompt, honchoPayload);
 
   if (!stream) {
     throw new Error('Failed to get stream');
